+++ src/dev/package.json
@@ -4,11 +4,11 @@
   "private": true,
   "dependencies": {
     "@eduzz/apps-toolbar-react": "latest",
-    "@eduzz/houston-experimental": "*",
-    "@eduzz/houston-forms": "*",
-    "@eduzz/houston-hooks": "*",
-    "@eduzz/houston-icons": "*",
-    "@eduzz/houston-ui": "*",
+    "@eduzz/houston-experimental": "0.52.0-beta.2",
+    "@eduzz/houston-forms": "0.52.0-beta.2",
+    "@eduzz/houston-hooks": "0.52.0-beta.2",
+    "@eduzz/houston-icons": "0.52.0-beta.2",
+    "@eduzz/houston-ui": "0.52.0-beta.2",
     "react": "^17",
     "react-dom": "^17",
     "react-router-dom": "^6.3.0",
