+++ src/pages/forms/locale.ts
@@ -1,70 +1,40 @@
-import { z, ZodIssueCode } from 'zod';
+/* eslint-disable no-template-curly-in-string */
+import { setLocale } from 'yup';
 
-z.setErrorMap((issue, ctx) => {
-  let message = ctx.defaultError;
-
-  switch (issue.code) {
-    case ZodIssueCode.invalid_type:
-      if (ctx.data === undefined || ctx.data === null) {
-        message = 'Deve ser preenchido';
-        break;
-      }
-
-      if (issue.expected === 'number') {
-        message = 'Deve ser um número válido';
-        break;
-      }
-
-      if (issue.expected === 'date') {
-        message = 'Deve ser uma data válida';
-        break;
-      }
-
-      message = 'Valor Inválido';
-      break;
-
-    case ZodIssueCode.invalid_string:
-      const invalidStringIssue = issue as z.ZodInvalidStringIssue;
-
-      if (invalidStringIssue.validation === 'email') {
-        message = 'Deve ser um e-mail válido';
-        break;
-      }
-
-      if (invalidStringIssue.validation === 'url') {
-        message = 'Deve ser uma URL válida';
-        break;
-      }
-
-      break;
-    case ZodIssueCode.too_small:
-      const toSmallIssue = issue as z.ZodTooSmallIssue;
-      message = `Deve ${toSmallIssue.type === 'number' ? 'ser' : 'ter'} no mínimo ${toSmallIssue.minimum}${
-        toSmallIssue.type === 'string'
-          ? ' caracteres'
-          : toSmallIssue.type === 'number'
-          ? ''
-          : toSmallIssue.minimum == 1
-          ? ' item'
-          : ' itens'
-      }`;
-      break;
-    case ZodIssueCode.too_big:
-      message = `Deve ter no máximo ${(issue as z.ZodTooBigIssue).maximum} caracteres`;
-      break;
-    case ZodIssueCode.invalid_date:
-      message = 'Deve ser uma data válida';
-      break;
-    case ZodIssueCode.invalid_literal:
-    case ZodIssueCode.invalid_union:
-    case ZodIssueCode.invalid_union_discriminator:
-    case ZodIssueCode.invalid_enum_value:
-    case ZodIssueCode.unrecognized_keys:
-    case ZodIssueCode.invalid_arguments:
-    case ZodIssueCode.invalid_return_type:
-    case ZodIssueCode.invalid_intersection_types:
-    case ZodIssueCode.not_multiple_of:
+setLocale({
+  mixed: {
+    default: 'Inválido',
+    required: 'Obrigatório',
+    oneOf: 'Deve ser um dos seguintes valores: ${values}',
+    notOneOf: 'Não pode ser um dos seguintes valores: ${values}',
+    notType: ({ type }) => (type === 'number' ? 'Deve ser um número' : 'Inválido')
+  },
+  string: {
+    length: 'Deve ter exatamente ${length} caracteres',
+    min: 'Deve ter pelo menos ${min} caracteres',
+    max: 'Deve ter no máximo ${max} caracteres',
+    email: 'E-mail inválido',
+    url: 'Deve ter um formato de URL válida',
+    trim: 'Não deve conter espaços no início ou no fim.',
+    lowercase: 'Deve estar em maiúsculo',
+    uppercase: 'Deve estar em minúsculo'
+  },
+  number: {
+    min: 'Deve ser no mínimo ${min}',
+    max: 'Deve ser no máximo ${max}',
+    lessThan: 'Deve ser menor que ${less}',
+    moreThan: 'Deve ser maior que ${more}',
+    // notEqual: 'Não pode ser igual à ${notEqual}',
+    positive: 'Deve ser um número positivo',
+    negative: 'Deve ser um número negativo',
+    integer: 'Deve ser um número inteiro'
+  },
+  date: {
+    min: 'Deve ser maior que a data ${min}',
+    max: 'Deve ser menor que a data ${max}'
+  },
+  array: {
+    min: 'Deve ter no mínimo ${min} itens',
+    max: 'Deve ter no máximo ${max} itens'
   }
-
-  return { message };
 });
